# p2-livecode-3-dbace23/docs/swagger.yaml
basePath: /
definitions:
  model.Activity:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      user_id:
        type: integer
    type: object
  model.CreatePostReq:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  model.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  model.Post:
    properties:
      author_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  model.RegisterReq:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
info:
  contact:
    email: halim.iskandar2323@gmail.com
    name: Halim Iskandar
  description: Small Instagram-like service (posts, likes, activities, users).
  title: Instagram Mini API
  version: "1.0"
paths:
  /v1/activities:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Activity'
            type: array
      security:
      - BearerAuth: []
      summary: My activities
      tags:
      - activities
  /v1/posts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Post'
            type: array
      security:
      - BearerAuth: []
      summary: List posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      parameters:
      - description: Create post payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreatePostReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create post
      tags:
      - posts
  /v1/posts/{id}:
    delete:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete post
      tags:
      - posts
    get:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Post detail
      tags:
      - posts
  /v1/users/login:
    post:
      consumes:
      - application/json
      description: Login with email + password, returns JWT
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - users
  /v1/users/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email validation
      parameters:
      - description: Register payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register user
      tags:
      - users
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Use:  Bearer <JWT>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"